pyproject.toml
[project]
name = "cohdd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core scientific computing
    "numpy>=1.26.0",
    "scipy>=1.10.0",
    "pandas>=1.5.0",
    "scikit-learn>=1.0.0",
    # Visualization
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    # llm control
    "litellm==1.67.2",
    "requests>=2.31.0",
    # AWS and cloud services
    "boto3>=1.26.0",
    "botocore>=1.29.0",
    # HTTP client
    "httpx>=0.28.1",
    # Environment and configuration
    "python-dotenv>=0.9.9",
    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-aws>=0.2.29",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-text-splitters>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain_tavily>=0.2.4",
    # Google ADK
    "google-adk",
    # LangGraph
    "langgraph>=0.5.4",
    "langgraph-cli>=0.3.5",
    # MCP (Model Context Protocol)
    "mcp>=1.12.1",
    "fastmcp>=0.1.0",
    "excel-mcp-server>=0.1.5",
    # Data processing and analysis
    "numba>=0.58.0",
    "dcor>=0.6.0",
    # Database and SQL
    "sqlalchemy>=2.0.0",
    # Visualization and mapping
    "plotly>=5.0.0",
    "kaleido>=0.2.1", # PNG export
    # Geocoding (free option)
    "geopy>=2.3.0",
    # Utilities
    "requests>=2.28.0",
    "requests-toolbelt>=1.0.0",
    "pydantic>=2.0.0",
    "tavily-python>=0.3.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "pytz>=2022.1",
    "packaging>=21.0",
    "joblib>=1.1.0",
    "networkx>=3.0",
    "orjson>=3.8.0",
    "tenacity>=8.0.0",
    "sniffio>=1.3.0",
    "h11>=0.14.0",
    "httpcore>=1.0.0",
    "idna>=3.4",
    "charset-normalizer>=3.0.0",
    "certifi>=2022.12.0",
    "jmespath>=1.0.1",
    "s3transfer>=0.6.0",
    "six>=1.16.0",
    "urllib3>=1.26.0",
    "annotated-types>=0.5.0",
    "anyio>=3.6.0",
    "jsonpatch>=1.32",
    "jsonpointer>=2.4",
    "kiwisolver>=1.4.4",
    "pyparsing>=3.0.9",
    "fonttools>=4.38.0",
    "cycler>=0.11.0",
    "contourpy>=1.0.6",
    "pillow>=9.5.0",
    "llvmlite>=0.39.0",
    "langsmith>=0.0.20",
    "tabulate>=0.9.0",
    "markdown>=3.8.2",
]
